+++ app/src/main/java/com/crrashh/sleepify/data/api/models/PointsHistoryItem.kt
@@ -3,7 +3,7 @@ package com.crrashh.sleepify.data.api.models
 import com.google.gson.annotations.SerializedName
 
 data class PointsHistoryItem(
-    val change: Int,
+    val change: Double,
     val reason: String,
     val operator: String?,
     @SerializedName("record_date") val recordDate: String
